+++ Changelog.md
@@ -8,6 +8,7 @@ CHANGELOG
 
 ### 3.9-preview
 
+- Add ContextMenuRequested routed event to ExplorerControlListView.
 - Update BrowsableObjectInfo related types in order to have more centralized plug-in handling.
 
 ### 3.8-preview